// sdk/nodejs/route53/enums.ts
// *** WARNING: this file was generated by the Pulumi Terraform Bridge (tfgen) Tool. ***
// *** Do not edit by hand unless you're certain you know what you are doing! ***

export const RecordType = {
    A: "A",
    AAAA: "AAAA",
    CNAME: "CNAME",
    CAA: "CAA",
    MX: "MX",
    NAPTR: "NAPTR",
    NS: "NS",
    PTR: "PTR",
    SOA: "SOA",
    SPF: "SPF",
    SRV: "SRV",
    TXT: "TXT",
} as const;

export type RecordType = (typeof RecordType)[keyof typeof RecordType];